requirements.txt
streamlit
agno
openai
yfinance
python-dotenv
feedparser
requests
